requirements3.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.5.0
aspy.yaml==1.3.0
Babel==2.7.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.7.1
blis==0.4.1
cachetools==3.1.0
catalogue==1.0.0
certifi==2019.3.9
cfgv==3.1.0
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.3
datasketch==1.4.3
distlib==0.3.1
Django==4.0.4
django-ajax-selects==2.2.0
django-braces==1.15.0
django-cors-headers==3.11.0
django-filter==1.0.4
django-ranged-response==0.2.0
django-rest-swagger==2.2.0
django-simple-captcha==0.5.17
djangorestframework==3.13.1
docutils==0.15.2
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
enum34==1.1.6
etherpad-lite==0.5
facebook-sdk==3.1.0
filelock==3.0.12
flake8==3.4.1
google-auth==1.6.3
identify==1.4.23
idna==2.8
imagesize==1.1.0
ipaddress==1.0.22
itypes==1.2.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.2
nodeenv==1.4.0
numpy==1.16.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.2
Pillow==6.2.1
plac==1.1.3
pre-commit==1.20.0
preshed==3.0.2
prometheus-client==0.6.0
psycopg2-binary==2.8.6
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2017.2
PyYAML==5.3.1
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
simplejson==3.17.2
six==1.12.0
snowballstemmer==2.0.0
soupsieve==1.9
spacy==2.3.2
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==1.0.2
thinc==7.4.1
toml==0.10.1
tqdm==4.48.0
tweepy==3.8.0
uritemplate==3.0.1
urllib3==1.24.2
virtualenv==20.0.27
wasabi==0.7.1
